refactor: Feature-First Architecture 에 따른 폴더 구조 개선 및 타입 파일명 수정
